% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_member_distance.R
\name{to_sparse_matrix}
\alias{to_sparse_matrix}
\title{edges is data.frame}
\usage{
to_sparse_matrix(edges)
}
\arguments{
\item{edges}{\code{data.frame} or a \verb{[sparseMatrix()]} with a \verb{$from} and \verb{$to} column with node ids.}
}
\value{
A sparse \code{Matrix} encoding the edges
}
\description{
edges is data.frame
}
